requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
atomicwrites==1.4.1
attrs==22.1.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==4.2.4
celery==5.0.5
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
convertdate==2.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cx-Oracle==8.1.0
distlib==0.3.6
Django==3.2
django-cors-headers==3.8.0
django-extensions==3.1.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.8.0
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
googlemaps==4.5.3
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.2
idna==3.2
importlib-metadata==4.0.1
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
joblib==1.1.0
kombu==5.0.2
lunardate==0.2.0
lxml==4.7.1
mariadb==1.0.8
MarkupSafe==2.0.1
mysqlclient==2.1.1
numpy==1.23.3
oauthlib==3.1.1
openapi-codec==1.3.2
openpyxl==3.0.9
outcome==1.2.0
packaging==21.0
pandas==1.4.4
paramiko==2.8.0
Pillow==8.4.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.18
protobuf==3.19.1
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyluach==1.3.0
PyMeeus==0.5.11
PyMySQL==1.0.2
PyNaCl==1.4.0
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypiwin32==223
pysftp==0.2.9
PySocks==1.7.1
pyTelegramBotAPI==4.5.1
pytest==6.2.5
python-dateutil==2.8.1
python-dotenv==0.19.1
pytz==2021.1
pywin32==302
PyYAML==6.0
reportlab==3.6.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.4.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.41
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
workalendar==16.3.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
zipp==3.8.1
